Added condition on the Evaluation Job to not delete the test pdf file in the resources.
Changed to ZonedDateTime to format date as ISO8601

+++ src/main/kotlin/org/ionproject/integration/application/job/ISELEvaluationsJob.kt
@@ -102,7 +102,8 @@ class ISELEvaluationsJob(
                 )
             }.orThrow()
         } finally {
-            File(path).delete()
+            if (!path.contains("test"))
+                File(path).delete()
         }
     }
